requirements.txt
pdfplumber
langchain-community
langchain-text-splitters
pypdf
openai
python-dotenv